+++ include/libp2p/protocol/gossip/gossip.hpp
@@ -64,10 +64,10 @@ namespace libp2p::protocol::gossip {
    public:
     virtual ~Gossip() = default;
 
-    /// Starts instance
+    /// Starts client and server
     virtual void start() = 0;
 
-    /// Stops instance
+    /// Stops client and server
     virtual void stop() = 0;
 
     /// Message received on subscription.
