+++ include/libp2p/protocol/gossip/impl/topic_subscriptions.hpp
@@ -1,0 +1,80 @@
+/**
+ * Copyright Soramitsu Co., Ltd. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef LIBP2P_PROTOCOL_GOSSIP_TOPIC_SUBSCRIPTIONS_HPP
+#define LIBP2P_PROTOCOL_GOSSIP_TOPIC_SUBSCRIPTIONS_HPP
+
+#include <deque>
+
+#include <libp2p/protocol/gossip/impl/peer_set.hpp>
+
+namespace libp2p::protocol::gossip {
+
+  class Connectivity;
+
+  /// Per-topic subscriptions
+  class TopicSubscriptions {
+   public:
+    /// Ctor. Dependencies are passed by ref because this object is a part of
+    /// RemoteSubscriptions and lives only within its scope
+    TopicSubscriptions(TopicId topic, const Config &config,
+                       Connectivity &connectivity);
+
+    /// Returns true if no peers subscribed and not self-subscribed and
+    /// no fanout period at the moment (empty item may be erased)
+    bool empty() const;
+
+    /// Forwards message to mesh members and announce to other subscribers
+    void onNewMessage(const TopicMessage::Ptr &msg, const MessageId &msg_id,
+                      bool is_published_locally, Time now);
+
+    /// Periodic job needed to update meshes and shift "I have" caches
+    void onHeartbeat(Time now);
+
+    /// Local host subscribes or unsubscribes, this affects mesh
+    void onSelfSubscribed(bool self_subscribed);
+
+    /// Remote peer subscribes to topic
+    void onPeerSubscribed(const PeerContextPtr &p);
+
+    /// Remote peer unsubscribes from topic
+    void onPeerUnsubscribed(const PeerContextPtr &p);
+
+    /// Remote peer includes this host into its mesh
+    void onGraft(const PeerContextPtr &p);
+
+    /// Remote peer kicks this host out of its mesh
+    void onPrune(const PeerContextPtr &p);
+
+   private:
+    /// Adds a peer to mesh
+    void addToMesh(const PeerContextPtr &p);
+
+    /// Removes a peer from mesh
+    void removeFromMesh(const PeerContextPtr &p);
+
+    const TopicId topic_;
+    const Config &config_;
+    Connectivity &connectivity_;
+
+    /// This host subscribed to this topic or not, this affects mesh behavior
+    bool self_subscribed_;
+
+    /// Fanout period allows for publishing from this host without subscribing
+    Time fanout_period_ends_;
+
+    /// Peers subscribed to this topic, but not mesh members
+    PeerSet subscribed_peers_;
+
+    /// Mesh members to whom messages are forwarded in push manner
+    PeerSet mesh_peers_;
+
+    /// "I have" notifications for new subscribers aka seen messages cache
+    std::deque<std::pair<Time, MessageId>> seen_cache_;
+  };
+
+}  // namespace libp2p::protocol::gossip
+
+#endif  // LIBP2P_PROTOCOL_GOSSIP_TOPIC_SUBSCRIPTIONS_HPP
